refacto: move options-list in shadowdom
