+++ index.html
@@ -6,8 +6,13 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>z-autocomplete</title>
     <link rel="stylesheet" href="./src/index.css" />
-    <link rel="stylesheet" href="./src/demo.css" />
     <script type="module" src="./src/index.ts"></script>
+
+    <!-- The important part-->
+    <link rel="stylesheet" href="./src/demo.css" />
+    <script type="module" src="./src/demo.ts"></script>
+    <!-- /The important part-->
+
     <link href="https://unpkg.com/prismjs@1/themes/prism-twilight.min.css" rel="stylesheet" />
     <script src="https://unpkg.com/prismjs@1/components/prism-core.min.js" defer></script>
     <script src="https://unpkg.com/prismjs@1/plugins/autoloader/prism-autoloader.min.js" defer></script>
@@ -89,9 +94,9 @@
                 <button data-z-autocomplete-clear>
                     &times;
                 </button>
-
-                <ul data-z-autocomplete-options hidden></ul>
             </z-autocomplete>
+
+            <hr>
 
             <button id="full-example-clear">
                 Clear from outside
@@ -108,8 +113,6 @@
     &lt;button data-z-autocomplete-clear&gt;
         &amp;times;
     &lt;/button&gt;
-
-    &lt;ul data-z-autocomplete-options hidden&gt;&lt;/ul&gt;
 &lt;/z-autocomplete&gt;
 
 
@@ -117,7 +120,10 @@
     Clear from outside
 &lt;/button&gt;</code></pre>
 
-            <pre><code class="language-js">type OptionItem = {
+            <pre><code class="language-js">import './z-autocomplete.ts'
+import type { ZAutocomplete } from './z-autocomplete.ts';
+
+type OptionItem = {
     id: number,
     name: string
 };
@@ -137,7 +143,7 @@ const AUTOCOMPLETE_DATA: OptionItem[] = [
     { id: 12, name: 'Option12' },
 ];
 
-const fullExampleEl = document.querySelector('#full-example-el') as ZAutocomplete<OptionItem>;
+const fullExampleEl = document.querySelector('#full-example-el') as ZAutocomplete&lt;OptionItem&gt;;
 const fullExampleClear = document.querySelector('#full-example-clear') as HTMLButtonElement;
 
 fullExampleEl.fetchData = async (inputValue, abortSignal) => {
@@ -176,7 +182,7 @@ fullExampleEl.dataToOption = (data) => {
 }
 
 fullExampleEl.addEventListener('autocomplete', (e: Event) => {
-    const event = e as CustomEvent<OptionItem | undefined>;
+    const event = e as CustomEvent&lt;OptionItem | undefined&gt;;
 
     if (event.detail?.id === 5) event.preventDefault();
