+++ src/demo.css
@@ -37,7 +37,7 @@ z-autocomplete#full-example-el button {
     box-shadow: none;
 }
 
-z-autocomplete#full-example-el ul {
+z-autocomplete#full-example-el::part(options-list) {
     z-index: 2; /* fx: prismjs has some annoying styling, this is not necessary in your code */
     position: absolute;
     inset: calc(100% - .25rem) 0 auto 0;
@@ -51,17 +51,18 @@ z-autocomplete#full-example-el ul {
     overflow-y: auto;
 }
 
-z-autocomplete#full-example-el ul li {
+z-autocomplete#full-example-el::part(options-item) {
     padding: 8px 15px;
 }
 
-z-autocomplete#full-example-el ul li:not([aria-disabled="true"]):hover,
-z-autocomplete#full-example-el ul li[aria-selected="true"] {
+z-autocomplete#full-example-el::part(options-item):hover,
+z-autocomplete#full-example-el::part(options-item options-item--active) {
     cursor: pointer;
     background-color: var(--autocomplete-selected);
 }
 
-z-autocomplete#full-example-el ul li[aria-disabled="true"] {
+z-autocomplete#full-example-el::part(options-item options-item--disabled),
+z-autocomplete#full-example-el::part(options-item options-item--disabled):hover {
     background-color: var(--clr-accent);
     color: var(--autocomplete-selected);
 }
